- API refinement:   * Java-API:     + made Transaction#req_list() public     -> refined the method to return a list of raw results which can be processed by the user one result by another     + let the Otp methods for reading also return an ErlangValue object     + reduced code duplication     + let the Connection class encapsulate the exceptions from the OtpConnection object and throw only ConnectionException objects     + made ErlangValue#convertToErlang() public     + ErlangValue: properly convert empty erlang strings in string lists and JSON objects     + refactoring of the PubSub class to use the ErlangValue class as return values     + ReplicatedDHT: let the number of deleted values be an integer instead of a long     + use more strict result type checking     + updated documentation   * Python-JSON-API:     + fixed the parsing of the JSON results in the PubSub class     + refactoring of process_result() to process_result_[read|write|commit]() just like in Java     + use more strict result type checking     + added several more options and checks to the command line interface of Scalaris.py

+++ java-api/test/de/zib/scalaris/ReplicatedDHTTest.java
@@ -109,7 +109,7 @@ public class ReplicatedDHTTest {
 
         try {
             for (int i = 0; i < testData.length; ++i) {
-                long deleted = rdht.delete(testTime + key + i);
+                int deleted = rdht.delete(testTime + key + i);
                 DeleteResult result = rdht.getLastDeleteResult();
                 assertEquals(0, deleted);
                 assertEquals(0, result.ok);
@@ -148,7 +148,7 @@ public class ReplicatedDHTTest {
             
             // now try to delete the data:
             for (int i = 0; i < testData.length; ++i) {
-                long deleted = rdht.delete(testTime + key + i);
+                int deleted = rdht.delete(testTime + key + i);
                 DeleteResult result = rdht.getLastDeleteResult();
                 assertEquals(4, deleted);
                 assertEquals(4, result.ok);
@@ -195,7 +195,7 @@ public class ReplicatedDHTTest {
             
             // now try to delete the data:
             for (int i = 0; i < testData.length; ++i) {
-                long deleted = rdht.delete(testTime + key + i);
+                int deleted = rdht.delete(testTime + key + i);
                 DeleteResult result = rdht.getLastDeleteResult();
                 assertEquals(4, deleted);
                 assertEquals(4, result.ok);
@@ -209,7 +209,7 @@ public class ReplicatedDHTTest {
             
             // now try to delete the data:
             for (int i = 0; i < testData.length; ++i) {
-                long deleted = rdht.delete(testTime + key + i);
+                int deleted = rdht.delete(testTime + key + i);
                 DeleteResult result = rdht.getLastDeleteResult();
                 assertEquals(4, deleted);
                 assertEquals(4, result.ok);
